feat(metadata-sidebar): pr comments
