+++ src/elements/content-sidebar/MetadataSidebarRedesign.scss
@@ -3,7 +3,6 @@
 @import '~@box/blueprint-web-assets/tokens/tokens.scss';
 
 .bcs-MetadataSidebarRedesign {
-    width: 340px;
     background-color: $gray-02;
     padding-inline: 10px;
 }
