+++ src/elements/content-sidebar/MetadataSidebarRedesign.tsx
@@ -2,15 +2,21 @@
  * @file Redesigned Metadata sidebar component
  * @author Box
  */
-import { MetadataEmptyState } from '@box/metadata-editor';
+import * as React from 'react';
+import { FormattedMessage } from 'react-intl';
 import flow from 'lodash/flow';
 import getProp from 'lodash/get';
-import * as React from 'react';
+import { MetadataEmptyState } from '@box/metadata-editor';
 import type { MessageDescriptor } from 'react-intl';
-import { FormattedMessage } from 'react-intl';
-import type { ElementsXhrError } from '../../common/types/api';
-import type { BoxItem } from '../../common/types/core';
-import type { MetadataEditor, MetadataTemplate } from '../../common/types/metadata';
+import messages from '../common/messages';
+import { normalizeTemplates } from '../../features/metadata-instance-editor/metadataUtil';
+import { isUserCorrectableError } from '../../utils/error';
+import { withAPIContext } from '../common/api-context';
+import { withErrorBoundary } from '../common/error-boundary';
+import { withLogger } from '../common/logger';
+import { mark } from '../../utils/performance';
+import { EVENT_JS_READY } from '../common/logger/constants';
+import API from '../../api';
 import {
     FIELD_IS_EXTERNALLY_OWNED,
     FIELD_PERMISSIONS,
@@ -18,16 +24,37 @@ import {
     IS_ERROR_DISPLAYED,
     ORIGIN_METADATA_SIDEBAR_REDESIGN,
 } from '../../constants';
-import { normalizeTemplates } from '../../features/metadata-instance-editor/metadataUtil';
-import { isUserCorrectableError } from '../../utils/error';
-import { mark } from '../../utils/performance';
-import { withAPIContext } from '../common/api-context';
-import { withErrorBoundary } from '../common/error-boundary';
-import { withLogger } from '../common/logger';
-import { EVENT_JS_READY } from '../common/logger/constants';
-import messages from '../common/messages';
+import type { ElementsXhrError, ErrorContextProps } from '../../common/types/api';
+import type { BoxItem } from '../../common/types/core';
+import type { MetadataEditor, MetadataTemplate } from '../../common/types/metadata';
+import type { WithLoggerProps } from '../../common/types/logging';
+
 import './MetadataSidebarRedesign.scss';
-import { Metadata, MetadataSidebarRedesignProps } from './MetadataSidebarRedesignTypes';
+
+interface ExternalProps {
+    isBoxAiSuggestionsFeatureEnabled: boolean;
+    isFeatureEnabled: boolean;
+    selectedTemplateKey?: string;
+    templateFilters?: Array<string> | string;
+}
+
+interface PropsWithoutContext extends ExternalProps {
+    elementId: string;
+    fileId: string;
+    hasSidebarInitialized?: boolean;
+}
+
+// For some reason, when changing this type to an interface, the MetadataSidebarRedesign component throws an error on the onError prop.
+export type MetadataSidebarRedesignProps = {
+    api: API;
+} & PropsWithoutContext &
+    ErrorContextProps &
+    WithLoggerProps;
+
+interface Metadata {
+    editors: Array<MetadataEditor>;
+    templates: Array<MetadataTemplate>;
+}
 
 const MARK_NAME_JS_READY = `${ORIGIN_METADATA_SIDEBAR_REDESIGN}_${EVENT_JS_READY}`;
 
@@ -106,7 +133,7 @@ const MetadataSidebarRedesign: React.FunctionComponent<MetadataSidebarRedesignPr
     const showEmptyState = showEditor && editors && editors.length === 0;
 
     return (
-        <div className="bcs-MetadataSidebarRedesign">
+        <div className="bcs-MetadataSidebarRedesign bcs-content">
             <h3>
                 <FormattedMessage {...messages.sidebarMetadataTitle} />
             </h3>
